Fixed Prisma & TS warnings
