+++ pages/api/profile.ts
@@ -1,7 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import * as helpers from "@/lib/helpers";
-import { PrismaClient } from "@prisma/client";
-const db = new PrismaClient();
+import { db, checkAuth, filterUserData, decode } from "@/lib/helpers";
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method === "POST") {
@@ -12,13 +10,13 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 }
 
 const controller = {
-  get: async (req, res) => {
+  get: async (req: NextApiRequest, res: NextApiResponse) => {
     try {
-      const auth = await helpers.checkAuth(req);
-      const id = await helpers.decode(req.headers.token);
+      const auth = await checkAuth(req);
+      const id = await decode(req.headers.token);
       if (auth) {
         const data = await db.user.findUnique({ where: { id: id } });
-        const filtered = await helpers.filterUserData(data);
+        const filtered = await filterUserData(data);
         res.status(200).json(filtered);
       } else {
         res.status(401).json({ error: "Unauthorised" });
@@ -28,10 +26,10 @@ const controller = {
       res.status(500).json({ error: "An error has occured" });
     }
   },
-  post: async (req, res) => {
+  post: async (req: NextApiRequest, res: NextApiResponse) => {
     try {
-      const auth = await helpers.checkAuth(req);
-      const id = await helpers.decode(req.headers.token);
+      const auth = await checkAuth(req);
+      const id = await decode(req.headers.token);
       if (auth) {
         const data = await db.user.update({
           where: {
